Fix the HRNet pretrained models paths
Despite the current HRNet pretrained models paths responds well and you can access the files, they have a sufix "?dl=0" that is useless and it generates problems because Windows don't allow the symbol '?' in the filenames and in linux there are some problems as well. Without that suffix (that I think is an error), it works fine. (¡¡muchas gracias por publicar esta librería, David, es un gran aporte a la comunidad!!)